waterclaw/requirements.txt
pyyaml>=6.0
requests>=2.28
numpy>=1.24
pandas>=2.0
